Configuration of stylelint lib and plugin (#4)
Stylelint is used to check css format.
I configured it to support also scss (Sass).
For some reason Stylelint vscode plugin is not recognizing the configuration file of the plugin. For that reason I had to include the same configurations of the .stylelintrc.json file to the .vscode/settings.json file.